refactor(ir): move TileBufSignature out of codegen into ir
TileBufSignature is a pure function of ir::TileType (no codegen
dependencies) but lived in namespace pypto::codegen under
include/pypto/codegen/pto/. Its only consumer is the IR pass
FoldNoOpReshape, which had to reverse-#include a codegen header -- the
same layering inversion (IR depends on codegen) as the IsBuiltinOp and
wrapper-direction cleanups.

Move the header to include/pypto/ir/transforms/utils/ and change the
namespace to pypto::ir; update FoldNoOpReshape's include and drop the
codegen:: qualifier. No codegen code referenced the type, so codegen is
unaffected. Behavior-preserving (header-only struct moved verbatim).
